move listdatum benchmark to pfs suite
